Refactor test fixture by moving fixtures only under the trainer scope to global fixture, and modified in test_abstract_trainer tests to use fixtures as opposed to rely on helper class import. This would benefit future test addition by exposing more useful fixtures
